input translation kinda doesn't suck now

+++ win/curses/cursmesg.c
@@ -477,7 +477,7 @@ done:
 						free(tmpstr);
 					} else {
 						/* clean up the leftovers on the next line, if we didn't scroll it away */
-						mvwprintw(win, my--, border_space, "%*c", strlen(linestarts[nlines]), ' ');
+						mvwprintw(win, my--, border_space, "%*c", (int)strlen(linestarts[nlines]), ' ');
 					}
 				}
 				break;
